Completed Shortcut and Added zoom
- Shortcut now working : shortcuts have to be added to method
getShortcut, by associating an Integer[] containing extendedKeyCode of
the Key, with a String representing command. If command doesn't exist in
actionPerformed, a window pop up error shows at runtime.

- Zoom features added : Zoom and Dezoom can be done with Ctrl/Wheel but
destroy the image with repeated applications. TODO : Resize the original
image instead of the resized one for each zoom and dezoom and applicate
a filter to avoid disaster with strong zooms/dezooms.